[CLIENT] Adapts tables in NodeForm and About to use Table component
