+++ src/client/src/content/forms/node/terminals/NodeFormTerminalTable.helpers.ts
