+++ src/client/src/content/forms/node/variants/FunctionNode.tsx
@@ -2,7 +2,7 @@ import { Aspect } from "@mimirorg/typelibrary-types";
 import { Control, UseFormRegister } from "react-hook-form";
 import { FormNodeLib } from "../../types/formNodeLib";
 import { NodeFormAttributes } from "../attributes/NodeFormAttributes";
-import { NodeFormTerminals } from "../terminals/NodeFormTerminals";
+import { NodeFormTerminalTable } from "../terminals/NodeFormTerminalTable";
 
 export interface FunctionNodeProps {
   control: Control<FormNodeLib>;
@@ -11,7 +11,7 @@ export interface FunctionNodeProps {
 
 export const FunctionNode = ({ control, register }: FunctionNodeProps) => (
   <>
-    <NodeFormTerminals control={control} />
+    <NodeFormTerminalTable control={control} />
     <NodeFormAttributes control={control} register={register} aspects={[Aspect.Function]} />
   </>
 );
