+++ src/client/src/content/home/components/about/components/terminal/TerminalTable.tsx
@@ -1,11 +1,13 @@
 import { ArrowLeft, ArrowRight, SwitchHorizontal } from "@styled-icons/heroicons-outline";
 import { useTheme } from "styled-components";
 import textResources from "../../../../../../assets/text/TextResources";
-import { Flexbox } from "../../../../../../complib/layouts";
+import { Table, Tbody, Td, Th, Thead, Tr } from "../../../../../../complib/data-display";
+import { Box, Flexbox } from "../../../../../../complib/layouts";
+import { Text } from "../../../../../../complib/text";
+import { useMediaQuery } from "../../../../../../hooks/useMediaQuery";
 import { TerminalItem } from "../../../../types/TerminalItem";
 import { AttributeInfoButton } from "../attribute/AttributeInfoButton";
 import { TerminalButton } from "./TerminalButton";
-import { TerminalTableContainer, TerminalTableData, TerminalTableHeader } from "./TerminalTable.styled";
 
 interface TerminalTableProps {
   terminals: TerminalItem[];
@@ -17,50 +19,69 @@ interface TerminalTableProps {
  * @param terminals to show inside the table
  * @constructor
  */
-export const TerminalTable = ({ terminals }: TerminalTableProps) => (
-  <TerminalTableContainer>
-    <thead>
-      <tr>
-        <TerminalTableHeader textAlign={"left"}>{textResources.TERMINAL_TABLE_NAME}</TerminalTableHeader>
-        <TerminalTableHeader textAlign={"left"}>{textResources.TERMINAL_TABLE_DIRECTION}</TerminalTableHeader>
-        <TerminalTableHeader textAlign={"center"}>{textResources.TERMINAL_TABLE_AMOUNT}</TerminalTableHeader>
-        <TerminalTableHeader textAlign={"left"}>{textResources.TERMINAL_TABLE_ATTRIBUTES}</TerminalTableHeader>
-      </tr>
-    </thead>
-    <tbody>
-      {terminals.map((x, i) => (
-        <TerminalTableRow key={i + x.name} {...x} />
-      ))}
-    </tbody>
-  </TerminalTableContainer>
-);
-
-const TerminalTableRow = ({ name, amount, color, direction, attributes }: TerminalItem) => {
+export const TerminalTable = ({ terminals }: TerminalTableProps) => {
   const theme = useTheme();
   const directionIconSize = 20;
+  const adjustAmountAlignment = useMediaQuery("screen and (min-width: 1500px)");
 
   return (
-    <tr>
-      <TerminalTableData textAlign={"left"}>
-        <Flexbox alignItems={"center"} gap={theme.tyle.spacing.base}>
-          <TerminalButton variant={"small"} as={"div"} color={color} direction={direction} />
-          {name}
-        </Flexbox>
-      </TerminalTableData>
-      <TerminalTableData textAlign={"left"}>
-        <Flexbox alignItems={"center"} gap={theme.tyle.spacing.base}>
-          {direction === "Input" && <ArrowRight size={directionIconSize} />}
-          {direction === "Output" && <ArrowLeft size={directionIconSize} />}
-          {direction === "Bidirectional" && <SwitchHorizontal size={directionIconSize} />}
-          {direction}
-        </Flexbox>
-      </TerminalTableData>
-      <TerminalTableData textAlign={"center"}>{amount}</TerminalTableData>
-      <TerminalTableData textAlign={"left"}>
-        <Flexbox flexWrap={"wrap"} gap={theme.tyle.spacing.base}>
-          {attributes && attributes.map((a, i) => <AttributeInfoButton key={i} {...a} />)}
-        </Flexbox>
-      </TerminalTableData>
-    </tr>
+    <Table borders width={"100%"}>
+      <Thead>
+        <Tr>
+          <Th>
+            <Text as={"span"} color={theme.tyle.color.sys.primary.base}>
+              {textResources.TERMINAL_TABLE_NAME}
+            </Text>
+          </Th>
+          <Th>
+            <Text as={"span"} color={theme.tyle.color.sys.primary.base}>
+              {textResources.TERMINAL_TABLE_DIRECTION}
+            </Text>
+          </Th>
+          <Th textAlign={adjustAmountAlignment ? "center" : "left"}>
+            <Text as={"span"} color={theme.tyle.color.sys.primary.base}>
+              {textResources.TERMINAL_TABLE_AMOUNT}
+            </Text>
+          </Th>
+          <Th>
+            <Text as={"span"} color={theme.tyle.color.sys.primary.base}>
+              {textResources.TERMINAL_TABLE_ATTRIBUTES}
+            </Text>
+          </Th>
+        </Tr>
+      </Thead>
+      <Tbody>
+        {terminals.map((terminal, i) => (
+          <Tr key={i + terminal.name}>
+            <Td data-label={textResources.TERMINAL_TABLE_NAME}>
+              <Flexbox alignItems={"center"} gap={theme.tyle.spacing.base}>
+                <TerminalButton variant={"small"} as={"div"} color={terminal.color} direction={terminal.direction} />
+                {terminal.name}
+              </Flexbox>
+            </Td>
+            <Td data-label={textResources.TERMINAL_TABLE_DIRECTION}>
+              <Flexbox alignItems={"center"} gap={theme.tyle.spacing.base}>
+                {terminal.direction === "Input" && <ArrowRight style={{ flexShrink: 0 }} size={directionIconSize} />}
+                {terminal.direction === "Output" && <ArrowLeft style={{ flexShrink: 0 }} size={directionIconSize} />}
+                {terminal.direction === "Bidirectional" && (
+                  <SwitchHorizontal style={{ flexShrink: 0 }} size={directionIconSize} />
+                )}
+                {terminal.direction}
+              </Flexbox>
+            </Td>
+            <Td data-label={textResources.TERMINAL_TABLE_AMOUNT} textAlign={adjustAmountAlignment ? "center" : "left"}>
+              {terminal.amount}
+            </Td>
+            <Td data-label={textResources.TERMINAL_TABLE_ATTRIBUTES}>
+              <Box display={"flex"} flexWrap={"wrap"} minWidth={"200px"} gap={theme.tyle.spacing.base}>
+                {terminal.attributes?.map((a, index) => (
+                  <AttributeInfoButton key={index} {...a} />
+                ))}
+              </Box>
+            </Td>
+          </Tr>
+        ))}
+      </Tbody>
+    </Table>
   );
 };
